Fix signed vs unsigned comparisons, found by GCC

+++ src/echo.c
@@ -204,7 +204,7 @@ veread(const char *fp, char *buf, size_t nbuf, int flag, va_list ap)
 		if (dynbuf) {
 			if ((buf = malloc(maclcur->l_used + 1)) == NULL)
 				return (NULL);
-		} else if (maclcur->l_used >= nbuf)
+		} else if ((size_t)maclcur->l_used >= nbuf)
 			return (NULL);
 		bcopy(maclcur->l_text, buf, maclcur->l_used);
 		buf[maclcur->l_used] = '\0';
@@ -334,7 +334,7 @@ veread(const char *fp, char *buf, size_t nbuf, int flag, va_list ap)
 			while ((y = kremove(i++)) >= 0 && y != '\n') {
 				int t;
 
-				if (dynbuf && epos + 1 >= nbuf) {
+				if (dynbuf && (size_t)(epos + 1) >= nbuf) {
 					void *newp;
 					size_t newsize = epos + epos + 16;
 					if ((newp = realloc(buf, newsize))
@@ -343,7 +343,7 @@ veread(const char *fp, char *buf, size_t nbuf, int flag, va_list ap)
 					buf = newp;
 					nbuf = newsize;
 				}
-				if (!dynbuf && epos + 1 >= nbuf) {
+				if (!dynbuf && (size_t)(epos + 1) >= nbuf) {
 					dobeep();
 					ewprintf("Line too long");
 					return (emptyval);
@@ -518,7 +518,7 @@ veread(const char *fp, char *buf, size_t nbuf, int flag, va_list ap)
 			c = getkey(FALSE);
 			/* fallthrough */
 		default:
-			if (dynbuf && epos + 1 >= nbuf) {
+			if (dynbuf && (size_t)(epos + 1) >= nbuf) {
 				void *newp;
 				size_t newsize = epos + epos + 16;
 				if ((newp = realloc(buf, newsize)) == NULL)
@@ -526,7 +526,7 @@ veread(const char *fp, char *buf, size_t nbuf, int flag, va_list ap)
 				buf = newp;
 				nbuf = newsize;
 			}
-			if (!dynbuf && epos + 1 >= nbuf) {
+			if (!dynbuf && (size_t)(epos + 1) >= nbuf) {
 				dobeep();
 				ewprintf("Line too long");
 				return (emptyval);
@@ -629,7 +629,7 @@ complt(int flags, int c, char *buf, size_t nbuf, int cpos, int *nx)
 		 */
 		if (nxtra < 0 && nhits > 1 && c == ' ')
 			nxtra = 1; /* ??? */
-		for (i = 0; i < nxtra && cpos < nbuf; ++i) {
+		for (i = 0; i < nxtra && (size_t)cpos < nbuf; ++i) {
 			buf[cpos] = lh2->l_name[cpos];
 			eputc(buf[cpos++]);
 		}
@@ -795,7 +795,7 @@ complt_list(int flags, char *buf, int cpos)
 			len = strlcat(linebuf, lh2->l_name + preflen,
 			    linesize);
 			width += maxwidth;
-			if (len < width && width < linesize) {
+			if (len < (size_t)width && (size_t)width < linesize) {
 				/* pad so the objects nicely line up */
 				memset(linebuf + len, ' ',
 				    maxwidth - strlen(lh2->l_name + preflen));
